Implement favorites functionality and refactor navigation
- Implement "Favorite" feature, including `ToggleFavorite` and `GetFavoritePokemon` use cases.
- Add `isFavorite` field to `Pokemon` domain model, `PokemonEntity`, and DTOs.
- Update `PokemonDao` with queries to update favorite status and fetch favorite Pokémon.
- Update `PokemonRepositoryImpl` to persist favorite status during data synchronization.
- Enhance `DetailsViewModel` and `DetailsTopBar` to handle favorite toggling with UI feedback.
- Update `HomeViewModel` to support filtering by favorites via search query ("favoritos" or "favorites").
- Refactor navigation by moving `Screen` and `NavHost` logic to a dedicated `NavGraph` file.
- Change `height` and `weight` types from `Int` to `Double` across models and mappers for better precision (decimeters/hectograms to meters/kilograms).
- Increment Room database version to 4.

+++ app/src/main/java/br/me/vitorcsouza/pokedex/data/remote/dto/PokemonDetailDto.kt
@@ -8,9 +8,9 @@ data class PokemonDetailDto(
     @SerializedName("name")
     val name: String,
     @SerializedName("height")
-    val height: Int,
+    val height: Double,
     @SerializedName("weight")
-    val weight: Int,
+    val weight: Double,
     @SerializedName("types")
     val types: List<TypeEntryDto>,
     @SerializedName("stats")
